requirements.txt
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
configparser==5.2.0
crayons==0.4.0
cryptography==36.0.1
cycler==0.11.0
Distance==0.1.3
dnspython==2.1.0
Flask==2.0.2
Flask-Cors==3.0.10
fonttools==4.27.1
h11==0.12.0
idna==3.3
imageio==2.13.5
itsdangerous==2.0.1
Jinja2==3.0.3
#joblib @ file:///tmp/build/80754af9/joblib_1635411271373/work
joblib
kiwisolver
MarkupSafe==2.0.1
matplotlib
#networkx @ file:///tmp/build/80754af9/networkx_1633639043937/work
networkx
nltk==3.6.7
numpy
outcome==1.1.0
packaging==21.3
pandas
Pillow==8.4.0
pycparser==2.21
pymongo==4.0.1
pyOpenSSL==21.0.0
pyparsing==3.0.6
pysummarize==0.6.0
python-dateutil==2.8.2
pytz==2021.3
regex==2021.11.10
requests==2.26.0
scipy
selenium
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
stopwords==1.0.0
#tqdm @ file:///tmp/build/80754af9/tqdm_1635330843403/work
tqdm
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.7
webdriver-manager==3.5.2
Werkzeug==2.0.2
wincertstore==0.2
wsproto==1.0.0
gunicorn
